catalogue sync moved into index.js. skus and prices are now being added to bpc
