+++ packages/scrapers/Dockerfile
@@ -1,0 +1,8 @@
+FROM node:18-alpine3.15
+
+WORKDIR .
+COPY ./ .
+
+RUN yarn
+
+CMD ["node", "index.js"]
